refactor: Refactor schemas part 2
