+++ src/components/dashboard-resource/EmbedModal.tsx
@@ -1,9 +1,8 @@
 import { CopyBlock, dracula } from 'react-code-blocks';
+import { Dashboard, Resource } from '@/schemas/configuration/configurationSchema';
 import { useLocale, useTranslations } from 'next-intl';
 
-import { Dashboard } from '@/schemas/configuration/dashboardsSchema';
 import { Modal } from 'react-bootstrap';
-import { Resource } from '@/schemas/configuration/configurationSchema';
 import { useState } from 'react';
 
 type EmbedModalProps =
