+++ src/configuration.ts
@@ -1,6 +1,5 @@
-import { Category, Configuration, Resource } from '@/schemas/configuration/configurationSchema';
+import { Category, Configuration, Dashboard, Resource } from '@/schemas/configuration/configurationSchema';
 
-import { Dashboard } from './schemas/configuration/dashboardsSchema';
 import YAML from 'yaml';
 import { promises as fs } from 'node:fs';
 import { merge } from 'ts-deepmerge';
