+++ src/schemas/configuration/configurationSchema.ts
@@ -1,114 +1,25 @@
-import { dashboardsSchema } from './dashboardsSchema';
+import {
+    axisPairSchema,
+    defaultFilterSchema,
+    embeddedResourceSchema,
+    geoJSONResourceSchema,
+    jsonResourceSchema,
+    resourceSchema,
+} from './resourcesSchema';
+import {
+    carouselContentSchema,
+    dashboardContentSchema,
+    dashboardContentSizeSchema,
+    dashboardSchema,
+    dashboardsSchema,
+    externalContentSchema,
+    externalLinkContentSchema,
+    internalLinkContentSchema,
+    resourceContentSchema,
+    textContentSchema,
+} from './dashboardsSchema';
+
 import { z } from 'zod';
-
-const baseResourceSchema = z
-    .object({
-        id: z.string(),
-        source: z.string().url(),
-        name: z.string(),
-        description: z.string().optional(),
-    })
-    .strict();
-
-export const embeddedResourceSchema = baseResourceSchema
-    .extend({
-        type: z.union([z.literal('HTML'), z.literal('PDF')]),
-    })
-    .strict();
-
-const axisPairSchema = z
-    .object({
-        xAxis: z.string(),
-        yAxis: z.string(),
-    })
-    .strict();
-
-const defaultFilterSchema = z.union([
-    z.string(),
-    z
-        .object({
-            min: z.number().transform(String).optional(),
-            max: z.number().transform(String).optional(),
-        })
-        .strict()
-        .refine((data) => data.min !== undefined || data.max !== undefined, {
-            message: "At least one of 'min' or 'max' must be defined",
-        }),
-]);
-
-const transformableResourceSchema = baseResourceSchema
-    .extend({
-        skipPropertiesRegEx: z.string().optional(),
-        renameProperties: z.record(z.string()).optional(),
-        numberFormat: z.union([z.literal('en'), z.literal('de')]).default('en'),
-    })
-    .strict();
-
-export const jsonResourceSchema = transformableResourceSchema
-    .extend({
-        type: z.union([z.literal('JSON'), z.literal('CSV')]),
-        defaultFilters: z.record(z.string(), defaultFilterSchema).optional(),
-        visualizations: z
-            .object({
-                barChart: z
-                    .object({
-                        axisPairs: z.array(axisPairSchema),
-                        layout: z.union([z.literal('horizontal'), z.literal('vertical')]).default('horizontal'),
-                    })
-                    .strict()
-                    .optional(),
-                table: z.record(z.never()).default({}),
-            })
-            .strict()
-            .default({ table: {} }),
-    })
-    .strict()
-    .refine(
-        (data) => {
-            if (data.skipPropertiesRegEx && data.renameProperties) {
-                const regex = new RegExp(data.skipPropertiesRegEx, 'u');
-                return !Object.keys(data.renameProperties).some((key) => regex.test(key));
-            }
-            return true;
-        },
-        {
-            message: 'skipPropertiesRegEx should not match any keys in renameProperties',
-            path: ['skipPropertiesRegEx'],
-        },
-    );
-
-export const geoJSONResourceSchema = transformableResourceSchema
-    .extend({
-        type: z.literal('GeoJSON'),
-        coordinateFormat: z.union([z.literal('LatLng'), z.literal('UTM')]).default('LatLng'),
-        visualizations: z
-            .object({
-                map: z
-                    .object({
-                        groupKey: z.string().optional(),
-                    })
-                    .strict()
-                    .default({}),
-            })
-            .strict()
-            .default({ map: {} }),
-    })
-    .strict()
-    .refine(
-        (data) => {
-            if (data.skipPropertiesRegEx && data.renameProperties) {
-                const regex = new RegExp(data.skipPropertiesRegEx, 'u');
-                return !Object.keys(data.renameProperties).some((key) => regex.test(key));
-            }
-            return true;
-        },
-        {
-            message: 'skipPropertiesRegEx should not match any keys in renameProperties',
-            path: ['skipPropertiesRegEx'],
-        },
-    );
-
-export const resourceSchema = z.union([jsonResourceSchema, geoJSONResourceSchema, embeddedResourceSchema]);
 
 const categoryBaseSchema = z
     .object({
@@ -146,8 +57,21 @@ export type EmbeddedResource = z.infer<typeof embeddedResourceSchema>;
 export type JSONResource = z.infer<typeof jsonResourceSchema>;
 export type GeoJSONResource = z.infer<typeof geoJSONResourceSchema>;
 export type AxisPair = z.infer<typeof axisPairSchema>;
-export type Resource = z.infer<typeof resourceSchema>;
-export type Configuration = z.infer<typeof configurationSchema>;
 export type Filter = z.infer<typeof defaultFilterSchema>;
+export type Resource = z.infer<typeof resourceSchema>;
+
 export type Category = z.infer<typeof categorySchema>;
+
+export type DashboardContentSize = z.infer<typeof dashboardContentSizeSchema>;
+export type DashboardContent = z.infer<typeof dashboardContentSchema>;
+export type DashboardResourceContent = z.infer<typeof resourceContentSchema>;
+export type DashboardTextContent = z.infer<typeof textContentSchema>;
+export type DashboardCarouselContent = z.infer<typeof carouselContentSchema>;
+export type DashboardExternalContent = z.infer<typeof externalContentSchema>;
+export type DashboardResourceLinkContent = z.infer<typeof internalLinkContentSchema>;
+export type DashboardExternalLinkContent = z.infer<typeof externalLinkContentSchema>;
+export type Dashboard = z.infer<typeof dashboardSchema>;
+
 export type Appearance = z.infer<typeof appearanceSchema>;
+
+export type Configuration = z.infer<typeof configurationSchema>;
